[REFACTOR] Phase 4: fix api data->procedures layering, split file storage (task 4.3)
- Move the upload-session store (Map, chunk parsing/assembly, session
  accessors) to api/src/shared/uploadSessions.ts; procedures/uploads.ts
  is now a thin RPC facade (cancelUpload + re-exports) and the files
  resource no longer imports upward into the procedures layer
- Split the on-disk binary half of files.ts (dir config + test hooks,
  atomic write, read/delete, path sanitization) into
  data/resources/fileStorage.ts; files.ts keeps resource ops, metadata
  SQL, and upload preparation

Pure file moves; verified api build, lint, 86/86 unit tests.
