+++ api/src/data/resources/files.ts
@@ -1,6 +1,3 @@
-import { mkdir, readFile, rename, unlink, writeFile } from "node:fs/promises";
-import { dirname, join, resolve } from "node:path";
-import { fileURLToPath } from "node:url";
 import { and, asc, eq, gt } from "drizzle-orm";
 
 import type {
@@ -26,7 +23,15 @@ import {
 	deleteUploadSession,
 	getUploadSession,
 	uploadSessionToBuffer,
-} from "../../procedures/uploads";
+} from "../../shared/uploadSessions";
+import {
+	deleteFileBinary,
+	deleteFileBinaryIfExists,
+	hasNodeErrorCode,
+	NODE_ENOENT,
+	readFileBinary,
+	writeFileBinary,
+} from "./fileStorage";
 
 // Types
 
@@ -35,31 +40,7 @@ type PreparedFileUpload = {
 	metadataPayload: DATA_EVY_File;
 };
 
-// Storage configuration
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
-let filesDir = resolve(join(__dirname, "..", "..", "public", "files"));
-let uploadTmpDir = resolve(join(__dirname, "..", "..", "public", "uploads"));
-
-// Test hooks
-
-export function setFileStorageDirsForTest(params: {
-	filesDir: string;
-	uploadTmpDir: string;
-}): void {
-	filesDir = params.filesDir;
-	uploadTmpDir = params.uploadTmpDir;
-}
-
-export function resetFileStorageDirsForTest(): void {
-	filesDir = resolve(join(__dirname, "..", "..", "public", "files"));
-	uploadTmpDir = resolve(join(__dirname, "..", "..", "public", "uploads"));
-}
-
 // Resource operations
-
-const NODE_ENOENT = "ENOENT" as const;
 
 export async function listFileRowsWithBinary(
 	db: EvyDb,
@@ -220,41 +201,6 @@ async function createFileFromUpload(params: {
 	};
 }
 
-// Binary storage
-
-// exported for tests
-export async function writeFileBinary(params: {
-	id: string;
-	bytes: Buffer;
-}): Promise<void> {
-	await mkdir(filesDir, { recursive: true });
-	await mkdir(uploadTmpDir, { recursive: true });
-
-	const finalPath = filePath(params.id);
-	const tmpPath = join(uploadTmpDir, `${sanitizeFileId(params.id)}.tmp`);
-
-	try {
-		await writeFile(tmpPath, params.bytes);
-		await rename(tmpPath, finalPath);
-	} catch (err) {
-		await deletePathIfExists(finalPath);
-		await deletePathIfExists(tmpPath);
-		throw err;
-	}
-}
-
-async function readFileBinary(id: string): Promise<Buffer> {
-	return readFile(filePath(id));
-}
-
-async function deleteFileBinary(id: string): Promise<void> {
-	await unlink(filePath(id));
-}
-
-async function deleteFileBinaryIfExists(id: string): Promise<void> {
-	await deletePathIfExists(filePath(id));
-}
-
 // Response mapping
 
 async function fileRowToGetFileResponse(
@@ -274,33 +220,4 @@ async function fileRowToGetFileResponse(
 		updatedAt: metadata.updatedAt,
 		dataBase64: fileData.toString("base64"),
 	};
-}
-
-// Local helpers
-
-function filePath(id: string): string {
-	return resolve(join(filesDir, sanitizeFileId(id)));
-}
-
-function sanitizeFileId(id: string): string {
-	return id.replace(/[^a-zA-Z0-9-]/g, "");
-}
-
-function hasNodeErrorCode(err: unknown, code: string): boolean {
-	return (
-		typeof err === "object" &&
-		err !== null &&
-		"code" in err &&
-		err.code === code
-	);
-}
-
-async function deletePathIfExists(path: string): Promise<void> {
-	try {
-		await unlink(path);
-	} catch (err) {
-		if (!hasNodeErrorCode(err, NODE_ENOENT)) {
-			throw err;
-		}
-	}
 }
